clean up file menu
